fixing previous commit
for some reason previous commit didn’t seem to be working, so forced
log string to delete itself before passing to req.log. Not sure if this
was the fix or not - doesn’t make sense - but it worked a few times
after this change so I’m leaving it in.